requirements.txt
# Core dependencies for production
ib_insync>=0.9.86
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Database
psycopg2-binary>=2.9.0

# Data sources
yfinance>=0.2.0
lxml>=4.9.0

# Progress bars
tqdm>=4.65.0

